refactor(console): make a module for the certification strategy impl
