+++ src/console/src/api/cdn.rs
@@ -7,7 +7,8 @@ use crate::cdn::proposals::{
     submit_proposal as make_submit_proposal,
 };
 use crate::cdn::strategies_impls::cdn::CdnHeap;
-use crate::cdn::strategies_impls::storage::{StorageCertificate, StorageState};
+use crate::cdn::strategies_impls::certification::StorageCertificate;
+use crate::cdn::strategies_impls::storage::StorageState;
 use crate::guards::caller_is_admin_controller;
 use crate::types::interface::DeleteProposalAssets;
 use ic_cdk_macros::{query, update};
@@ -20,7 +21,6 @@ use junobuild_shared::ic::call::ManualReply;
 use junobuild_shared::ic::UnwrapOrTrap;
 use junobuild_shared::types::core::DomainName;
 use junobuild_shared::types::domain::CustomDomains;
-
 // ---------------------------------------------------------
 // Proposal
 // ---------------------------------------------------------
