+++ src/console/src/api/http.rs
@@ -1,4 +1,5 @@
-use crate::cdn::strategies_impls::storage::{StorageCertificate, StorageState};
+use crate::cdn::strategies_impls::certification::StorageCertificate;
+use crate::cdn::strategies_impls::storage::StorageState;
 use ic_cdk_macros::query;
 use junobuild_storage::http::types::{
     HttpRequest, HttpResponse, StreamingCallbackHttpResponse, StreamingCallbackToken,
