+++ src/console/src/cdn/strategies_impls/mod.rs
@@ -1,2 +1,3 @@
 pub mod cdn;
+pub mod certification;
 pub mod storage;
