+++ src/console/src/cdn/strategies_impls/storage.rs
@@ -6,10 +6,7 @@ use crate::cdn::helpers::stable::{
     get_asset_stable, insert_asset_encoding_stable, insert_asset_stable,
 };
 use crate::cdn::storage::init_certified_assets;
-use crate::certification::cert::update_certified_data;
 use candid::Principal;
-use ic_certification::HashTree;
-use junobuild_auth::runtime::pruned_labeled_sigs_root_hash_tree;
 use junobuild_cdn::storage::errors::{
     JUNO_CDN_STORAGE_ERROR_CANNOT_GET_ASSET_UNKNOWN_REFERENCE_ID,
     JUNO_CDN_STORAGE_ERROR_CANNOT_INSERT_ASSET_ENCODING_UNKNOWN_REFERENCE_ID,
@@ -23,8 +20,7 @@ use junobuild_shared::types::core::Blob;
 use junobuild_shared::types::domain::CustomDomains;
 use junobuild_shared::types::state::Controllers;
 use junobuild_storage::strategies::{
-    StorageAssertionsStrategy, StorageCertificateStrategy, StorageStateStrategy,
-    StorageUploadStrategy,
+    StorageAssertionsStrategy, StorageStateStrategy, StorageUploadStrategy,
 };
 use junobuild_storage::types::config::StorageConfig;
 use junobuild_storage::types::state::FullPath;
@@ -253,20 +249,5 @@ impl StorageUploadStrategy for StorageUpload {
             }
             None => Err(JUNO_CDN_STORAGE_ERROR_CANNOT_GET_ASSET_UNKNOWN_REFERENCE_ID.to_string()),
         }
-    }
-}
-
-pub struct StorageCertificate;
-
-impl StorageCertificateStrategy for StorageCertificate {
-    fn update_certified_data(&self) {
-        update_certified_data();
-    }
-
-    // We use this function to access the auth signatures root hash
-    // in the storage crate when we build an HTTP response because the
-    // tree is a fork of both assets and signatures.
-    fn get_pruned_labeled_sigs_root_hash_tree(&self) -> HashTree {
-        pruned_labeled_sigs_root_hash_tree()
     }
 }
